chore: complete resolve of lint errors across all staged modules

+++ src/lib/api/schemas/products.ts
@@ -23,7 +23,7 @@ export const IndustryTypeEnum = z.enum([
 export const createProductSchema = z.object({
   name: z.string().min(3, 'El nombre debe tener al menos 3 caracteres'),
   description: z.string().optional(),
-  price: z.number().positive('El precio debe ser positivo'),
+  price: z.number().int().positive('El precio debe ser positivo'),
   stock: z.number().int().min(0, 'El stock no puede ser negativo').default(0),
   threshold_low: z.number().int().min(1).default(10),
   threshold_critical: z.number().int().min(1).default(3),
